Move common 'patching' behaviour out for reuse
To enable the patching of both the Specialist Publisher and Rummager schemas we need to share the behaviour to write to a give file. We could change to include modules and move away from inheritance but I feel it is still (just about) easy to read and work with.
