+++ lib/dfid-transition/patch/specialist_publisher/base.rb
@@ -1,15 +1,10 @@
+require 'dfid-transition/patch/patcher'
 require 'json'
 
 module DfidTransition
   module Patch
     module SpecialistPublisher
-      class Base
-        attr_reader :location
-
-        def initialize(location = nil)
-          @location = location || relative_path
-        end
-
+      class Base < Patcher
         def facet(name)
           schema_hash['facets'].find { |f| f['key'] == name } ||
             raise(KeyError.new("No #{name} facet found"))
@@ -19,21 +14,7 @@ module DfidTransition
           raise NotImplementedError
         end
 
-        def run
-          File.exist?(location) || raise(Errno::ENOENT.new(location))
-
-          mutate_schema
-
-          File.open(location, 'w+') do |target|
-            target.write(JSON.pretty_generate(schema_hash))
-          end
-        end
-
       private
-
-        def schema_hash
-          @schema_hash ||= JSON.parse(File.read(location))
-        end
 
         def relative_path
           File.expand_path(
